+++ src/HEADERS/input_baryons.h
@@ -6,15 +6,15 @@
 #define INPUT_BARYONS_H
 
 /**
-   @fn int baryon_contractions( struct baryon_info *baryons , int *nbaryons , const struct inputs *INPUT , const int nprops , const GLU_bool first_pass ) 
+   @fn int baryon_contractions( struct baryon_info *baryons , size_t *nbaryons , const struct inputs *INPUT , const size_t nprops , const GLU_bool first_pass ) 
    @brief baryon contraction information
    @return #SUCCESS or #FAILURE
  */
 int
 baryon_contractions( struct baryon_info *baryons , 
-		     int *nbaryons ,
+		     size_t *nbaryons ,
 		     const struct inputs *INPUT ,
-		     const int nprops ,
+		     const size_t nprops ,
 		     const GLU_bool first_pass ) ;
 
 #endif
