+++ src/HEADERS/input_mesons.h
@@ -6,15 +6,15 @@
 #define INPUT_MESONS_H
 
 /**
-   @fn int meson_contractions( struct meson_info *mesons , int *nmesons , const struct inputs *INPUT , const int nprops , const GLU_bool first_pass )
+   @fn int meson_contractions( struct meson_info *mesons , size_t *nmesons , const struct inputs *INPUT , const size_t nprops , const GLU_bool first_pass )
    @brief meson contraction information, packs mesons struct
    @return #SUCCESS or #FAILURE
  */
 int
 meson_contractions( struct meson_info *mesons , 
-		    int *nmesons ,
+		    size_t *nmesons ,
 		    const struct inputs *INPUT ,
-		    const int nprops ,
+		    const size_t nprops ,
 		    const GLU_bool first_pass ) ;
 
 #endif
