+++ src/HEADERS/input_tetras.h
@@ -6,15 +6,15 @@
 #define INPUT_TETRAS_H
 
 /**
-   @fn int tetra_contractions( struct tetra_info *tetras , int *ntetras , struct inputs *INPUT , const int nprops , const GLU_bool first_pass )
+   @fn int tetra_contractions( struct tetra_info *tetras , size_t *ntetras , struct inputs *INPUT , const size_t nprops , const GLU_bool first_pass )
    @brief get the tetraquark information from the input file
    @return #SUCCESS or #FAILURE
  */
 int
 tetra_contractions( struct tetra_info *tetras , 
-		    int *ntetras ,
+		    size_t *ntetras ,
 		    struct inputs *INPUT ,
-		    const int nprops ,
+		    const size_t nprops ,
 		    const GLU_bool first_pass ) ;
 
 #endif
