+++ src/HEADERS/read_config.h
@@ -52,7 +52,7 @@ get_config_SUNC( FILE *__restrict CONFIG ,
 		 const struct head_data HEAD_DATA ) ;
 
 /**
-   @fn struct site* read_gauge_file( struct head_data *HEAD_DATA , const char *config_in , const int refdims[ ND ] )
+   @fn struct site* read_gauge_file( struct head_data *HEAD_DATA , const char *config_in , const size_t refdims[ ND ] )
    @brief read the gauge field in
    @param HEAD_DATA :: header data stripped out
    @param config_in :: gauge configuration file
@@ -61,6 +61,6 @@ get_config_SUNC( FILE *__restrict CONFIG ,
 struct site*
 read_gauge_file( struct head_data *HEAD_DATA , 
 		 const char *config_in ,
-		 const int refdims[ ND ] ) ;
+		 const size_t refdims[ ND ] ) ;
 
 #endif
