+++ src/IO/input_WME.c
@@ -10,8 +10,8 @@
 static int
 matrix_element_tokens( struct WME_info *wme ,
 		       const char *WME_str ,
-		       const int nprops ,
-		       const int meas_idx ) 
+		       const size_t nprops ,
+		       const size_t meas_idx ) 
 {
   printf( "\n" ) ;
 
@@ -32,17 +32,17 @@ matrix_element_tokens( struct WME_info *wme ,
   if( get_contraction_map( &( wme -> map[3] ) , token , nprops ) == FAILURE ) {
     return FAILURE ;
   }
-  printf( "[IO] WME_%d :: Contracting prop %d with prop %d with %d with %d \n" , 
+  printf( "[IO] WME_%zu :: Contracting prop %zu with prop %zu with %zu with %zu \n" , 
 	  meas_idx , wme -> map[0] , wme -> map[1] , wme -> map[2] , wme -> map[3] ) ;
 
   // output file
   if( ( token = (char*)strtok( NULL , "," ) ) == NULL ) return unexpected_NULL( ) ;
   sprintf( wme -> outfile , "%s" , token ) ;
-  printf( "[IO] WME_%d :: Contraction file in %s \n" , meas_idx , token ) ;
+  printf( "[IO] WME_%zu :: Contraction file in %s \n" , meas_idx , token ) ;
 
   // tell us if we get more than we expect
   if( ( token = (char*)strtok( NULL , "," ) ) != NULL ) {
-    printf( "[IO] WME_%d :: Unexpected extra contraction info %s \n" ,
+    printf( "[IO] WME_%zu :: Unexpected extra contraction info %s \n" ,
 	    meas_idx , token ) ;
   }
   return SUCCESS ;
@@ -51,15 +51,15 @@ matrix_element_tokens( struct WME_info *wme ,
 // matrix elemen contractions expect prop map of 4 indices
 int
 matrix_element_contractions( struct WME_info *wme , 
-			     int *nWME ,
+			     size_t *nWME ,
 			     const struct inputs *INPUT ,
-			     const int nprops ,
+			     const size_t nprops ,
 			     const GLU_bool first_pass ) 
 {
   *nWME = 0 ;
   char str[ 32 ] ;
   while( *nWME < MAX_CONTRACTIONS ) {
-    sprintf( str , "WME%d" , *nWME ) ;
+    sprintf( str , "WME%zu" , *nWME ) ;
     const int WME_idx = tag_search( str ) ;
     if( WME_idx == FAILURE ) return SUCCESS ;
     if( first_pass == GLU_FALSE ) {
