+++ src/IO/input_baryons.c
@@ -10,8 +10,8 @@
 static int
 threepoint_tokens( struct baryon_info *baryons ,
 		   const char *baryon_str ,
-		   const int nprops ,
-		   const int meas_idx ,
+		   const size_t nprops ,
+		   const size_t meas_idx ,
 		   const char *message ) 
 {
   printf( "\n" ) ;
@@ -29,18 +29,19 @@ threepoint_tokens( struct baryon_info *baryons ,
   if( get_contraction_map( &( baryons -> map[2] ) , token , nprops ) == FAILURE ) {
     return FAILURE ;
   }
-  printf( "[IO] %s_%d :: Contracting prop %d with prop %d with prop %d \n" , 
-	  message , meas_idx , baryons -> map[0] , baryons -> map[1] , baryons -> map[2] ) ;
+  printf( "[IO] %s_%zu :: Contracting prop %zu with prop %zu with prop %zu \n" , 
+	  message , meas_idx , baryons -> map[0] , 
+	  baryons -> map[1] , baryons -> map[2] ) ;
 
   // output file
   if( ( token = (char*)strtok( NULL , "," ) ) == NULL ) return unexpected_NULL( ) ;
   sprintf( baryons -> outfile , "%s" , token ) ;
-  printf( "[IO] %s_%d :: Contraction file in %s \n" , 
+  printf( "[IO] %s_%zu :: Contraction file in %s \n" , 
 	  message , meas_idx , token ) ;
 
   // tell us if we get more than we expect
   if( ( token = (char*)strtok( NULL , "," ) ) != NULL ) {
-    printf( "[IO] %s_%d :: Unexpected extra contraction info %s \n" ,
+    printf( "[IO] %s_%zu :: Unexpected extra contraction info %s \n" ,
 	    message , meas_idx , token ) ;
   }
   return SUCCESS ;
@@ -49,15 +50,15 @@ threepoint_tokens( struct baryon_info *baryons ,
 // baryon contractions of the form prop_idx1,prop_idx2,proptype,outfile
 int
 baryon_contractions( struct baryon_info *baryons , 
-		     int *nbaryons ,
+		     size_t *nbaryons ,
 		     const struct inputs *INPUT ,
-		     const int nprops ,
+		     const size_t nprops ,
 		     const GLU_bool first_pass ) 
 {
   *nbaryons = 0 ;
   char str[ 32 ] ;
   while( *nbaryons < MAX_CONTRACTIONS ) {
-    sprintf( str , "BARYON%d" , *nbaryons ) ;
+    sprintf( str , "BARYON%zu" , *nbaryons ) ;
     const int baryon_idx = tag_search( str ) ;
     if( baryon_idx == FAILURE ) return SUCCESS ;
     if( first_pass == GLU_FALSE ) {
