+++ src/IO/input_tetras.c
@@ -10,8 +10,8 @@
 static int
 tetra_tokens( struct tetra_info *tetras ,
 	      const char *tetra_str ,
-	      const int nprops ,
-	      const int meas_idx ,
+	      const size_t nprops ,
+	      const size_t meas_idx ,
 	      const char *message ) 
 {
   printf( "\n" ) ;
@@ -28,18 +28,18 @@ tetra_tokens( struct tetra_info *tetras ,
   if( get_contraction_map( &( tetras -> map[2] ) , token , nprops ) == FAILURE ) {
     return FAILURE ;
   }
-  printf( "[IO] %s_%d :: Contracting prop %d with prop %d with prop %d\n" , 
+  printf( "[IO] %s_%zu :: Contracting prop %zu with prop %zu with prop %zu\n" , 
 	  message , meas_idx , tetras -> map[0] , tetras -> map[1] , tetras -> map[2] ) ;
 
   // output file
   if( ( token = (char*)strtok( NULL , "," ) ) == NULL ) return unexpected_NULL( ) ;
   sprintf( tetras -> outfile , "%s" , token ) ;
-  printf( "[IO] %s_%d :: Contraction file in %s \n" , 
+  printf( "[IO] %s_%zu :: Contraction file in %s \n" , 
 	  message , meas_idx , token ) ;
 
   // tell us if we get more than we expect
   if( ( token = (char*)strtok( NULL , "," ) ) != NULL ) {
-    printf( "[IO] %s_%d :: Unexpected extra contraction info %s \n" ,
+    printf( "[IO] %s_%zu :: Unexpected extra contraction info %s \n" ,
 	    message , meas_idx , token ) ;
   }
   return SUCCESS ;
@@ -48,15 +48,15 @@ tetra_tokens( struct tetra_info *tetras ,
 // tetra expects four contraction indices
 int
 tetra_contractions( struct tetra_info *tetras , 
-		    int *ntetras ,
+		    size_t *ntetras ,
 		    struct inputs *INPUT ,
-		    const int nprops ,
+		    const size_t nprops ,
 		    const GLU_bool first_pass ) 
 {
   *ntetras = 0 ;
   char str[ 32 ] ;
   while( *ntetras < MAX_CONTRACTIONS ) {
-    sprintf( str , "TETRA%d" , *ntetras ) ;
+    sprintf( str , "TETRA%zu" , *ntetras ) ;
     const int tetra_idx = tag_search( str ) ;
     if( tetra_idx == FAILURE ) return SUCCESS ;
     if( first_pass == GLU_FALSE ) {
